requirements.txt
configparser==3.5.0
coverage==4.3.4
Django==1.8.17
django-debug-toolbar==1.4
django-dotenv==1.4.1
django-kronos==1.0.0
django-model-utils==2.6
djangorestframework==3.6.2
enum34==1.1.6
flake8==3.2.1
mccabe==0.5.2
mock==2.0.0
oauthlib==2.0.1
psycopg2==2.6.2
pycodestyle==2.2.0
pyflakes==1.3.0
PyJWT==1.4.2
pytz==2016.10
python-openid==2.2.5
requests==2.13.0
requests-oauthlib==0.7.0
six==1.10.0
social-auth-app-django==1.0.0
social-auth-core==1.0.1
sqlparse==0.1.19
